refactor: mudando o nome do arquivo de logger
